### 2022-05-28 520387a
add license file for JOSS
