### 2022-05-28 #4436 @jeffreyhanson
Hi,

Thank you for considering my submission!

To address editorialbot's checks, I've just added in a LICENSE file (see [here](https://github.com/prioritizr/wdpar/commit/520387ad763a314dc772cfad4684ea415ecb52a3)).

Also, in terms of potential reviewers, I would suggest that individuals employed by UNEP-WCMC to work on the Protected Planet databases could provide especially insightful reviews, such as Ed Lewis (Github username: EdwardMLewis) or Cristina Lázaro (Github username CristinaLS). I would also suggest that individuals who have previously posted questions or bug reports on the software issue tracker could also provide useful reviews (assuming that this would not constitute a conflict of interest), such as Russell J. Gray (Github username RussellGrayxd), Phil Dyer (GitHub username PhDyellow), Johannes Schielein (GitHub username Jo-Schie), François Guilhaumon (GitHub username fguilhaumon), Jonas Geschke (GitHub username JonasGeschke), Jocelyne Shimin Sze (GitHub username  JocelyneSze), and Iacopo Neri (Github username neriiacopo). From the list of reviewers suggested by editorialbot, I would suggest the following individuals because R is among their preffered programming languages and they are interested in conservation: hvoltbb, struckma, DrMattG, and ghislainv.

Please let me know if there's any further details I can provide?
