### 2022-06-02 #4436 @martinfleis
Hi @jeffreyhanson,

I happy to edit this submission. On the first look, I noticed that it very small in scope and the actual R code has only 1285 lines, which is a borderline size for the JOSS. See https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort for details.

I have now flagged the submission for a scope check by the wider editorial team to figure out if it is large enough to be considered as substantial scholarly effort. Give us two weeks to get back to you please.
