### 2022-06-26 #4436 @jeffreyhanson
Hi @martinfleis, it's been several weeks now, so I just wanted to follow up and ask if there's anything I can do to help? Is there any further information I can provide to help with determining if the submission meets the journal's critieria?
