+++ NEWS.md
@@ -1,3 +1,8 @@
+# wdpar 1.3.2.4
+
+- Fix compatibility with GEOS (version 3.11.0).
+- Fix broken URLs in documentation.
+
 # wdpar 1.3.2.3
 
 - Fix typos in documentation, add JOSS paper, and add contributing guide.
