### 2022-07-08 ad8b7b1
Fix GEOS (fix #50, fix #51) (#52)
fix broken urls too
