+++ cran-comments.md
@@ -1,6 +1,6 @@
 Dear CRAN volunteers,
 
-Thank you for reviewing this submission. The main updates in this version include (i) fixing compatibility issues with the Protected Planet website to enable users to download data, (ii) improving data cleaning procedures, (iii) updating code and documentation to reflect the recent addition of the World Database on Other Effective Area-Based Conservation Measures (WDOECM) to the Protected Planet website, and (iv) removing unused package dependencies. The package now also includes the prepr R package as an optional dependency to augment data processing (https://github.com/dickoa/prepr).
+Thank you for reviewing this submission. The main updates in this version include (i) making data download procedures more robust, (ii) fixing compatibility issues with recent updates to GEOS (version 3.11.0), (iii) improving the documentation.
 
 Cheers,
 
@@ -19,7 +19,7 @@ Jeff
 
 0 errors | 0 warnings | 2 note
 
-# CRAN check notes for new submission
+# Package check notes
 
 * Possibly misspelled words in DESCRIPTION:
   WDOECM (7:28)
@@ -30,13 +30,6 @@ Jeff
   prepr
 
   **The prepr R package is an optional dependency that is available on GitHub (<https://github.com/dickoa/prepr>). Instructions for installing the prepr R package are provided in the DESCRIPTION file and the package README file.**
-
-# CRAN check notes for current version on CRAN
-
-* Namespace in Imports field not imported from: ‘withr’
-  All declared Imports should be used.
-
-  **This note has been addressed by removing the withr R package from the Imports field.**
 
 # Downstream dependencies
 
