+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: wdpar
 Type: Package
-Version: 1.3.2.4
+Version: 1.3.3
 Title: Interface to the World Database on Protected Areas
 Description: Fetch and clean data from the World Database on Protected
     Areas (WDPA) and the World Database on Other Effective Area-Based
