### 2022-07-19 #4436 @Jo-Schie
Sure. I'd used and I am still using this software. It's a great package and I'd be happy to contribute in form of a review.

Just to repeat what is expected for me: basically downloading, installing, testing and cross-checking with the criteria mentioned [here](JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html)... right?

Are there any requirements in terms of style or length of the review @martinfleis ?
