### 2022-07-19 #4436 @martinfleis
Thanks @Jo-Schie! I will formally start the review once I get a second reviewer and then the review issue contains all the information again but the review criteria cover most of it. We ask reviewers to follow [this checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) which will be generated for you by our bot in the review issue later.
