### 2022-07-19 #4594 @editorialbot
[REVIEW]: wdpar: Interface to the World Database on Protected Areas
**Submitting author:** @jeffreyhanson (<a href="http://orcid.org/0000-0002-4716-6134">Jeffrey Hanson</a>)
**Repository:** https://github.com/prioritizr/wdpar
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.3.2
