### 2022-07-23 #4594 @Jo-Schie
Check installation for:

- [ ] Ubuntu/Linux
- [ ] MacOS
- [ ] Windows @DrMattG : Do you have a Windows machine where you could test installation? @martinfleis : do we both need to test installation on all platforms or can we have a division of labor somehow? I work mostly with linux but I'm also able to check on MacOS.
