### 2022-07-23 #4594 @DrMattG
I have windows - so can test installation
