### 2022-07-23 prioritizr/wdpar#53 @Jo-Schie
JOSS Review: Improve documentation on geo-processing steps and its effects on the original geometries
Hi @jeffreyhanson .

I think it could be helpful for users to extend a little the documentation on the internal geo-processing steps and what they actually do to the data (geometries). This should be part of the vignette, but also of the paper (which so far very much focuses very much on describing the need, but not the internals of the package).

From my first test, there are a few things that should be discussed:

- [x] Overview of the applied geo-processing steps in `wdpa_clean`. -> maybe this could be also a graphical representation.
- [x] which geo-processing steps from `wdpa_clean` affect the original geometries?
- [x] How do they affect the geometries and what are consequences for the use (e.g. question of scale)?
- [x] How do default settings affect geometries (discuss that default is country scale analysis)?
- [x] How can we fine-tune settings e.g. to work on different scales and what are the effects on the geometries?
- [x] How are overlapping polygon boundaries resolved and what are the consequences for analysis -> e.g. if there are two overlapping areas with different IUCN categories, how is the boundary resolved and what are the consequences for calculating area statistics.

If possible, I would recommend that you use interactive maps in your vignette instead of the map that you plotted (which does not allow to really see anything). Below you can find some sample code and some screenshots that might allow users to understand more intuitively what is done with the data. You could use and maybe extent this a little bit.

```R
# ----- reproduce quickstart tutorial from https://github.com/prioritizr/wdpar -----
# load packages
library(wdpar)
library(dplyr)
library(ggmap)

# download protected area data for Malta
mlt_raw_pa_data <- wdpa_fetch("Malta",
                              wait = TRUE,
                              download_dir = rappdirs::user_data_dir("wdpar"))

# clean Malta data
mlt_pa_data <- wdpa_clean(mlt_raw_pa_data)

# reproject data to longitude/latitude for plotting
mlt_pa_data <- st_transform(mlt_pa_data, 4326)

# download basemap imagery
bg <- get_stamenmap(unname(st_bbox(mlt_pa_data)),
                    zoom = 8,
                    maptype = "watercolor",
                    force = TRUE)

# make map
ggmap(bg) +
  geom_sf(aes(fill = IUCN_CAT), data = mlt_pa_data, inherit.aes = FALSE) +
  theme(axis.title = element_blank(), legend.position = "bottom")

# ----- interactive map to compare raw to processed data -----
library(mapview)

mapView(mlt_pa_data) + mapView(mlt_raw_pa_data, col.regions = "red")

# ----- compare higher precision processing -----
# problem: Data with the default setting gets quite distorted on local level
# (see mapview before)

# clean Malta data with higher geomtry precision
mlt_pa_data_highprecision <- wdpa_clean(mlt_raw_pa_data,
                                        geometry_precision = 10000)

mapView(mlt_pa_data) +
  mapView(mlt_raw_pa_data, col.regions = "red") +
  mapView(mlt_pa_data_highprecision, col.regions = "green")

# clean Malta data with higher geomtry precision and keep overlaps
mlt_pa_data_highprecision_overlap <- wdpa_clean(mlt_raw_pa_data,
                                                geometry_precision = 10000,
                                                erase_overlaps = FALSE)

mapView(mlt_pa_data) +
  mapView(mlt_raw_pa_data, col.regions = "red") +
  mapView(mlt_pa_data_highprecision, col.regions = "green")+
  mapView(mlt_pa_data_highprecision_overlap, col.regions = "purple")

```

**Original Data**
![image](https://user-images.githubusercontent.com/25028884/180598804-537b00ba-c1c0-4862-b180-16c88e968724.png)

**Default Settings**

**Higher Geometry precision**

**HIgher geometry precision and overlaps**

[Link](https://github.com/openjournals/joss-reviews/issues/4594) to review
