### 2022-07-23 #4594 @Jo-Schie
> @Jo-Schie If you have a possibility to check on more OS it would be welcome to do so, but don't feel obliged to test on all.

AFAIK all packages on CRAN are going through some automated installation checks anyways before publication...not sure if this is cross-Plattform though.
