### 2022-07-23 prioritizr/wdpar#54 @Jo-Schie
JOSS Review: Improve Statement of need / description of use-cases
It seems to me, that the statement of need in the paper and/or the introduction part in the documentation could be a little clearer.

My understanding is this: The `wdpar` package can help users of the `wdpa` and `wdoecm` to work more effectively with the data, in a broad sense by allowing user automated download and preprocessing according to scientific recommendations. You talk a lot about "data cleaning procedures" which are difficult to implement, but the link to the final use-cases is still a bit blurry from my perspective (you do though describe them a bit in the "research application" part.)

For me, the question arises whether `wdpar` is "only" intended to serve as a planning and reporting tool (i.e. display and quantification of protected area coverage & reducing over-estimations).

_-> I guess this is/was the main concern when you developed the package..._

or whether `wdpar` could also serve other "final" purposes, such as monitoring and evaluation of protected areas' effectiveness.

_-> this is probably what others have turned it into by using it for their own types of analysis_

Is this interpretation correct? If so, then it could make sense to inform users about that because the `wdpa_clean` function with its default settings is really focussed on area estimations.

Maybe you could write something along the line of:  _the primary intent of the package is to calculate protected area coverage to e.g. report on the Aichi policy progress, nevertheless, part of the downloading and preprocessing routines can also be helpful in other contexts such as spatial analysis of protected area effectiveness. If the final use-case is different from official PA coverage reporting, then users should be aware of the settings in the `wdpa_clean` function and fine-tune them to their needs. For monitoring, it might make sense e.g. to "retain a specific PA status", to "not exclude Unesco sites" or to "increase the geometrical precision" of the data-cleaning functions to have better area estimations at local level._

[Link](https://github.com/openjournals/joss-reviews/issues/4594) to the review threat.
