### 2022-07-25 prioritizr/wdpar#53 @jeffreyhanson
Thank you very much for these suggestions @Jo-Schie! I'll have a think about this and begin incorperating your feedback later this week.
