### 2022-07-25 prioritizr/wdpar#54 @jeffreyhanson
Yeah - this an excellent point - thank you! I'll update the manuscript to incorperate your suggestion.
