+++ .github/workflows/documentation.yaml
@@ -37,44 +37,41 @@ jobs:
     steps:
       - uses: actions/checkout@v2
 
-      - uses: r-lib/actions/setup-r@v1
+      - uses: r-lib/actions/setup-pandoc@v2
+
+      - uses: r-lib/actions/setup-r@v2
         with:
           r-version: ${{ matrix.config.r }}
-
-      - uses: r-lib/actions/setup-pandoc@v1
-
-      - name: Query dependencies
-        run: |
-          install.packages('remotes')
-          saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
-          writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
-        shell: Rscript {0}
-
-      - name: Cache R packages
-        if: runner.os != 'Windows'
-        uses: actions/cache@v2
-        with:
-          path: ${{ env.R_LIBS_USER }}
-          key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
-          restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
+          use-public-rspm: true
 
       - name: Install system dependencies
         run: |
-          sudo apt-get -y install \
-            libcurl4-openssl-dev coinor-libsymphony-dev coinor-libcbc-dev coinor-libclp-dev \
-            libudunits2-dev libgdal-dev libgeos-dev libproj-dev libglpk-dev
-          while read -r cmd
-          do
-            eval sudo $cmd
-          done < <(Rscript -e 'writeLines(remotes::system_requirements("ubuntu", "20.04"))')
+          sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
+          sudo apt update
+          sudo apt-get install \
+            libgdal-dev libgeos-dev libproj-dev \
+            libgmp3-dev libmpfr-dev libudunits2-dev
 
-      - name: Install dependencies
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          cache-version: 4
+          extra-packages: |
+            any::covr
+            any::rcmdcheck
+            any::XML
+            prepr=?ignore
+          needs: |
+            check
+            coverage
+
+      - name: Install spatial R packages
         run: |
-          remotes::install_deps(dependencies = TRUE)
-          remotes::install_cran("rcmdcheck")
-          remotes::install_cran("covr")
-          remotes::install_bioc("lpsymphony")
-          remotes::install_cran("urlchecker")
+          options(repos = "https://cloud.r-project.org/")
+          install.packages("sf")
+          install.packages("terra")
+          install.packages("lwgeom")
+          install.packages("rgdal")
+          remotes::install_github("dickoa/prepr")
         shell: Rscript {0}
 
       - name: Session information
@@ -83,11 +80,6 @@ jobs:
           pkgs <- installed.packages()[, "Package"]
           sessioninfo::session_info(pkgs, include_base = TRUE)
         shell: Rscript {0}
-
-      - name: Set environmental variables for package checks
-        if: matrix.config.pkgs == 'depends only'
-        run: |
-          echo "_R_CHECK_DEPENDS_ONLY_=true" >> $GITHUB_ENV
 
       - name: Check URLs
         if: runner.os == 'Linux' && matrix.config.r == 'release'
