+++ .github/workflows/R-CMD-check-ubuntu.yaml
@@ -61,6 +61,7 @@ jobs:
             any::covr
             any::rcmdcheck
             any::XML
+            any::remotes
             prepr=?ignore
           needs: |
             check
