+++ .github/workflows/documentation.yaml
@@ -59,6 +59,7 @@ jobs:
             any::covr
             any::rcmdcheck
             any::XML
+            any::remotes
             prepr=?ignore
           needs: |
             check
