+++ .github/workflows/R-CMD-check-ubuntu.yaml
@@ -61,7 +61,6 @@ jobs:
             any::covr
             any::rcmdcheck
             any::XML
-            any::remotes
             prepr=?ignore
           needs: |
             check
@@ -74,7 +73,7 @@ jobs:
           install.packages("terra")
           install.packages("lwgeom")
           install.packages("rgdal")
-          remotes::install_github("dickoa/prepr")
+          pak::pkg_install("dickoa/prepr")
         shell: Rscript {0}
 
       - name: Session information
