### 2022-08-04 3228f8e
remove unused terra from ci
