+++ .github/workflows/R-CMD-check-windows.yaml
@@ -65,7 +65,6 @@ jobs:
         run: |
           options(repos = "https://cloud.r-project.org/")
           install.packages("sf")
-          install.packages("terra")
           install.packages("lwgeom")
           install.packages("rgdal")
           install.packages("remotes")
