+++ .github/workflows/R-CMD-check-ubuntu.yaml
@@ -46,14 +46,6 @@ jobs:
           r-version: ${{ matrix.config.r }}
           use-public-rspm: true
 
-      - name: Install system dependencies
-        run: |
-          sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
-          sudo apt update
-          sudo apt-get install \
-            libgdal-dev libgeos-dev libproj-dev \
-            libgmp3-dev libmpfr-dev libudunits2-dev
-
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 4
@@ -68,11 +60,6 @@ jobs:
 
       - name: Install spatial R packages
         run: |
-          options(repos = "https://cloud.r-project.org/")
-          install.packages("sf")
-          install.packages("terra")
-          install.packages("lwgeom")
-          install.packages("rgdal")
           pak::pkg_install("dickoa/prepr")
         shell: Rscript {0}
 
