+++ .github/workflows/documentation.yaml
@@ -56,21 +56,15 @@ jobs:
         with:
           cache-version: 4
           extra-packages: |
-            any::covr
             any::rcmdcheck
             any::XML
+            any::urlchecker
             prepr=?ignore
           needs: |
             check
-            coverage
 
       - name: Install spatial R packages
         run: |
-          options(repos = "https://cloud.r-project.org/")
-          install.packages("sf")
-          install.packages("terra")
-          install.packages("lwgeom")
-          install.packages("rgdal")
           pak::pkg_install("dickoa/prepr")
         shell: Rscript {0}
 
