+++ .github/workflows/R-CMD-check-windows.yaml
@@ -63,7 +63,8 @@ jobs:
 
       - name: Install spatial R packages
         run: |
-          pak::pkg_install("dickoa/prepr")
+          install.packages("remotes")
+          remotes::install_github("dickoa/prepr")
         shell: Rscript {0}
 
       - name: Session information
