### 2022-08-08 ab9fbc4
attempt fix #54, - add in more information on customizations and associated applications
