### 2022-08-08 prioritizr/wdpar#54 @jeffreyhanson
@Jo-Schie, I've submitted a PR which aims to incorperate your feedback. When you get a chance, could you please take a look (see #54)?
