+++ paper/paper.Rmd
@@ -30,9 +30,13 @@ The \texttt{wdpar} \texttt{R} package provides an interface to data available on
 
 Area-based conservation measures are crucial for safeguarding biodiversity [@r5; @r6]. Examples of such measures include protected areas, marine reserves, and other effective area-based conservation measures (OECMs). Protected Planet is a key resource for area-based conservation measures, providing the World Database on Protected Areas (WDPA) and the World Database on Other Effective Area-Based Conservation Measures (WDOECM) [@r9]. These publicly available databases contain standardized data for over 270,000 protected areas and over 700 OECMs worldwide [@r9]. By detailing the designation, establishment, management, and spatial boundaries of area-based conservation measures [@r21], these databases play a vital role in monitoring and prioritizing conservation efforts [@r4; @r8].
 
-The WDPA and WDOECM require data cleaning procedures to prepare them for analysis [@r4; @r10]. These procedures include repairing invalid geometries in spatial boundaries, excluding areas that have yet to be fully implemented, excluding areas that are no longer designated, excluding UNESCO Biosphere Reserves [@r1], buffering areas represented by point localities [@r3], and removing spatial overlaps [@r2]. These procedures are critical to ensure that assessments of area-based conservation measures do not overestimate the spatial extent of such measures and their ability to conserve biodiversity [@r1; @r2]. Although these procedures are critical, they can be technically challenging to implement. The \texttt{wdpar} \texttt{R} package provides automated methods to complete these procedures following best practices [@r4; @r10]. Using the package, data cleaning procedures can applied without specialized knowledge, customized to particular use cases, and across the entire WDPA and WDOECM. As such, the \texttt{wdpar} \texttt{R} package helps increase accessibility to the databases.
+The WDPA and WDOECM require data cleaning procedures to prepare them for analysis [@r4; @r10]. These procedures include excluding areas that have yet to be fully implemented, areas that are no longer designated, and UNESCO Biosphere Reserves [@r1]. They also include geoprocessing procedures, such as repairing invalid geometries in spatial boundaries, buffering areas represented by point localities [@r3], and removing spatial overlaps [@r2]. Specifically, overlapping geometries are erased such that areas associated with more effective management categories or have historical precedence are retained. These procedures are critical to ensure accuracy in assessments of area-based conservation measures [@r1; @r2]. Yet, despite their importance, these procedures can be challenging to implement.
 
-# Research applications
+The \texttt{wdpar} \texttt{R} package provides automated methods to obtain and clean the WDPA and WDOECM. The data cleaning procedures implemented in the package follow best practices [@r4; @r10] and can be performed without specialized knowledge, customized to particular use cases, and applied to the global scale. By providing this functionality, the package aims to increase accessibility to the WDPA and WDOECM.
+
+# Applications
+
+The \texttt{wdpar} \texttt{R} package is designed to provide a reproducible tool for downloading and cleaning the WDPA and WDOECM. Although the default settings for the data cleaning procedures are well-suited for national scale reporting of protected area coverage, they can be customized for other applications. For example, the precision of spatial data processing procedures can be increased so that they are suitable for local scale analyses. This is especially important because the default precision may remove smooth edges at fine scales. Additionally, the data cleaning procedures can be customized to retain protected areas regardless of their status which, in turn, could be useful for monitoring and evaluation of protected area effectiveness.
 
[…]
