### 2022-08-08 prioritizr/wdpar#54 @jeffreyhanson
@Jo-Schie, I've submitted a PR which aims to incorperate feedback on both your issues (see #57). Specifically, this PR adds the following paragraph to address your comments. This paragraph makes it clear that the package has defaults which are designed specifically for national-scale protected area coverage assesments, and that these defaults can be customized (if desired) for other purposes. Thanks for proving an example paragraph as a starting point - I found that really helpful! How does that sound?

> The \texttt{wdpar} \texttt{R} package is designed to provide a reproducible tool for downloading and cleaning the WDPA and WDOECM. Although the default settings for the data cleaning procedures are well-suited for national scale reporting of protected area coverage, they can be customized for other applications. For example, the precision of spatial data processing procedures can be increased so that they are suitable for local scale analyses. This is especially important because the default precision may remove smooth edges at fine scales. Additionally, the data cleaning procedures can be customized to retain protected areas regardless of their status which, in turn, could be useful for monitoring and evaluation of protected area effectiveness.
