+++ paper/paper.bib
@@ -212,3 +212,14 @@
   publisher={Cambridge, UK: UNEP-WCMC},
   url={http://wcmc.io/WDPA_Manual}
 }
+
+@article{r22,
+  title={Protected areas and global conservation of migratory birds},
+  author={Runge, Claire A and Watson, James E M and Butchart, Stuart  HM and Hanson, Jeffrey O and Possingham, Hugh P and Fuller, Richard A},
+  journal={Science},
+  volume={350},
+  number={6265},
+  pages={1255--1258},
+  year={2015},
+  doi={10.1126/science.aac9180}
+}
