+++ vignettes/wdpar.Rmd
@@ -55,7 +55,7 @@ mlt_raw_pa_data <- wdpa_fetch(
 )
 ```
 
-Next, we will clean the data set. Briefly, the cleaning steps include: excluding protected areas that are not yet implemented, excluding protected areas with limited conservation value, replacing missing data codes (e.g. `"0"`) with missing data values (i.e. `NA`), replacing protected areas represented as points with circular protected areas that correspond to their reported extent, repairing any topological issues with the geometries, and erasing overlapping areas. Please note that, by default, spatial data processing is performed at a scale suitable for national scale analyses. If you require data suitable for fine scale analysis, please a higher level of precision when cleaning the data (e.g., using the augment: `geometry_precision = 10000`). For more information, see `wdpa_clean()`.
+Next, we will clean the data set. Briefly, the cleaning steps include: excluding protected areas that are not yet implemented, excluding protected areas with limited conservation value, replacing missing data codes (e.g. `"0"`) with missing data values (i.e. `NA`), replacing protected areas represented as points with circular protected areas that correspond to their reported extent, repairing any topological issues with the geometries, and erasing overlapping areas. Please note that, by default, spatial data processing is performed at a scale suitable for national scale analyses (see below for recommendations for local scale analyses). For more information on the data cleaning procedures, see `wdpa_clean()`.
 
 ```{r}
 # clean Malta data
@@ -221,6 +221,65 @@ print(pa_data)
 # calculate total area
 print(st_area(pa_data))
 ```
+
+### Recommended practices for local scale analyses
+
+The default parameters for the data cleaning procedures are well suited for national-scale analyses. Although these parameters reduce memory requirements and the time needed to complete the data cleaning procedures, they can produce protected area boundaries that appear overly "blocky" -- lacking smooth edges -- when viewed at finer scales. As such, it is strongly recommended to increase the level of spatial precision when cleaning data for local scale analyses (via the `geometry_precision` parameter of the `wdpa_clean()` function).
+
+Here we will explore the consequences of using the default parameters for the data cleaning procedures when working at a local scale. This will help illustrate why it can be important to adjust the spatial precision of the data cleaning procedures. To begin with, we will obtain data for a small protected area. Specifically, we will extract a protected area from the Malta dataset we downloaded earlier.
+
+```{r}
+# find id for smallest reserve in cleaned dataset
+mlt_reserve_id <- mlt_pa_data$WDPAID[which.min(mlt_pa_data$AREA_KM2)]
+
+# extract the smallest reserve from the raw dataset
+mlt_raw_reserve_data <-
+  mlt_raw_pa_data %>%
+  filter(WDPAID == mlt_reserve_id)
+
+# preview data
+print(mlt_raw_reserve_data)
+
+# visualize data
+plot(mlt_raw_reserve_data[, 1])
+```
+
+```{r, include = FALSE}
+assertthat::assert_that(
+  nrow(mlt_raw_reserve_data) == 1,
+  msg = "failed to extract small reserve from Malta data"
+)
+```
+
+We can see that the boundary for this protected area has a high level of detail. This suggests that the protected area data is available at a resolution that is sufficient to permit local scale analyses. To help understand the consequences of cleaning data with the default parameters, we will clean this dataset using the default parameters.
+
+```{r}
+# clean the data with default parameters
+mlt_default_cleaned_reserve_data <- wdpa_clean(mlt_raw_reserve_data)
+
+# preview data
+print(mlt_default_cleaned_reserve_data)
+
+# visualize data
+plot(mlt_default_cleaned_reserve_data[, 1])
+```
+
[…]
