### 2022-08-23 prioritizr/wdpar#54 @jeffreyhanson
@Jo-Schie, I just wanted to follow up and ask if the PR address your concerns?
