### 2022-08-23 prioritizr/wdpar#53 @jeffreyhanson
@Jo-Schie, I just wanted to follow up and ask if the PR address your concerns on this issue?
