### 2022-08-31 prioritizr/wdpar#53 @Jo-Schie
Hi @jeffreyhanson . I did not really realize that you'd already worked on this. I put it on my todo list and will come back asap.
