### 2022-08-31 prioritizr/wdpar#53 @jeffreyhanson
No worries - thanks! Please let me know if anything's not clear and I'll get back to you as soon as I can.
