### 2022-09-03 #4594 @jeffreyhanson
Sorry, I forgot to respond about the operating systems checks. @Jo-Schie is correct - CRAN does ineeded run automated checks. Briefly, these checks include installation checks, a range of code and documentation checks, and unit tests. They are run on several platforms including Debian, Windows, Fedora, and macOS (including on M1 Macs) (see here for platforms: https://cran.r-project.org/web/checks/check_flavors.html). As you can see here (https://cran.r-project.org/web/checks/check_results_wdpar.html), the package passes all of these checks across all platforms.
