### 2022-09-05 prioritizr/wdpar#59 @DrMattG
JOSS Review warning about out of date local data
Warning message:
In wdpa_fetch("Norway", wait = TRUE, download_dir = rappdirs::user_data_dir("wdpar")) :
  local data is out of date: jul 2022
(https://github.com/prioritizr/wdpar/blob/40e8e79501b7caf16f684ec0f386a6299b241d07/R/wdpa_fetch.R#L183)
What does it mean if the local data is out of date? Do I need to worry about this warning? It would be good to have some comment on this in the documentation (or somewhere).
https://github.com/openjournals/joss-reviews/issues/4594
