### 2022-09-05 prioritizr/wdpar#59 @jeffreyhanson
The package will by default load the latest version of the database that is available on your computer (e.g., because you previously downloaded the data). This warning message is telling you that this version is now out of date, and a newer version is available on online. I guess you would need to worry about this warning if you are running analyses that require the latest version, but generally this isn't the case. Sure, I can add some info to the docs on this message.
