### 2022-09-12 prioritizr/wdpar#53 @jeffreyhanson
@Jo-Schie, I just wanted to follow up and ask if you'd a chance to go through the PR?
