### 2022-09-19 prioritizr/wdpar#59 @jeffreyhanson
Sorry for the slow progress on this @DrMattG. Just to keep you in the loop, I'm planning on creating a PR to incorperate this feedback - once I've merged the PR to address the issues raised by @Jo-Schie. Since I'm really bad at dealing with merge conflicts, I'd rather not create multiple branches at the same time and then have to juggle conflicts when merging them. Did you have any other issues/concerns (apart from #58 and #59) as part of your review? Or will I have addressed all your major concerns once these are addressed? Thanks for your patience!
