### 2022-09-19 prioritizr/wdpar#53 @jeffreyhanson
@Jo-Schie, it's been over a month since I originally created the PR to address your feedback, is there anything I can do to help make it easier for you to review it?
