### 2022-09-21 prioritizr/wdpar#54 @jeffreyhanson
@Jo-Schie, if you get a moment, could you please check whether the PR addresses this issue too?
