### 2022-09-21 prioritizr/wdpar#54 @Jo-Schie
yep. I am on it. Is there any link to the new manuscript of the paper? I can only find the old one.
