### 2022-09-21 prioritizr/wdpar#54 @jeffreyhanson
Brillian - thank you! Yeah, here's the .Rmd file (https://github.com/prioritizr/wdpar/blob/joschie-review/paper/paper.Rmd) and the PDF version (https://github.com/prioritizr/wdpar/blob/joschie-review/paper/paper.pdf).
