### 2022-09-21 prioritizr/wdpar#54 @Jo-Schie
okay. thanks. One question. I do have the feeling that the manuscript is very short, and you do neither show a little snippet of the code of the package nor do you show any output of an example application. I have checked a couple of paper in JOSS who all do that see e.g. [here](https://joss.theoj.org/papers/10.21105/joss.04740), [here](https://joss.theoj.org/papers/10.21105/joss.04015) or [here](https://joss.theoj.org/papers/10.21105/joss.04303) for papers published just lately.

Now I do understand, that showing code and outputs is not a strict precondition to publish the paper, so I am happy to close the issue also as is. Nevertheless, I suggest that you could add a little bit of the code snippets from your quick-start and maybe a plot to it. I think it could improve the overall quality of the paper since readers will already see what they can expect, but it is up to you if you want to do that or not.

so just give it a thought and comment back so I know when to close the issue.
