### 2022-09-21 prioritizr/wdpar#61 @Jo-Schie
JOSS Review: Add small POC about Performance claims 
JOSS [Review criteria from the thread](https://github.com/openjournals/joss-reviews/issues/4594) ask to assess whether there are any performance claims of the package that can be assessed. In your quick-start vignette, you state that:

_"The wdpar R package can be used to clean large datasets assuming that sufficient computational resources and time are available. Indeed, it can clean data spanning large countries, multiple countries, and even the full global datatset (sic). When processing the full global dataset, it is recommended to use a computer system with at least 32 GB RAM available and to allow for at least one full day for the data cleaning procedures to complete"_

Is it possible for you to provide a very short example that sustains that claim? I thought maybe you could randomly sample e.g. 500 areas from the global dataset and process them recording starttime and stoptime and report back. This would help to extrapolate and sustain the claim. I can also try to create a small example, but I never worked on the global data, so I guess it might be easier for you to do so.

I would not care too much about the efficiency of a geospatial software if it was designed for local scale analysis but the beauty of the `wdpar` package is, at least in theory, that one could create summary stats on the whole global progress towards achieving the AICHI targets (or the post-AICHI targets if e.g. the 30 by 30 goal is going to be approved by the global conservation community, - btw. something you could also mention in your use case description). So in that specific case, I would really like to see whether one could use the package for that.
