### 2022-09-21 prioritizr/wdpar#54 @jeffreyhanson
Awesome - thank you very much! Yeah, I wasn't sure if papers were supposed to include coding examples or not. Yeah, I can put in a small demo and include a figure.
