### 2022-09-21 prioritizr/wdpar#62 @Jo-Schie
JOSS Review: Add statement about state of the field 
Last issue from my side. [JOSS Review criteria ask i](https://github.com/openjournals/joss-reviews/issues/4594)f the author describes the _"State of the field: Do the authors describe how this software compares to other commonly-used packages?"_

I guess that the `wdpar` package is pretty unique in the sense that there is no other R (or Python) package AFAIK that do exactly the same or something very similar to what you do. So it could be informative to the reader to write a short sentence in your manuscript that tells them that probably this is the only piece of software to deal with this very specific issue. I always enjoy if an R-package also lists alternatives s that could serve me and how they are different... so knowing that basically at the moment there is no alternative to the `wdpar` package could be a shortcut for some readers to not investigate too much into finding other alternatives (I guess there is also no other software outside the Python / R world to do so.) + it increases the value of your package to the reader in the sense that it is unique in what it does.

Nevertheless, there is e.g. the Digital Observatory for Protected Areas [(DOPA)](https://dopa.jrc.ec.europa.eu/dopa/) which provides country summary stats on protected area progress towards AICHI. You could cite them because probably they have to apply exactly the same methods as you do and do allow you to create summary stats to a limited degree.
