### 2022-09-21 prioritizr/wdpar#61 @Jo-Schie
Sorry. Maybe I oversaw it...but how many Polygons did you process? This was not the global WDPA right?
