### 2022-09-21 prioritizr/wdpar#61 @Jo-Schie
Ah okay. Got it. It is the global data but without unionizing. Hummm. I thought your section on big data and processing overnight was also referring to unionizing ie ereasing overlaps. **Did you ever do that for the global data?**

 If so, could there be a method for more or less benchmarking it? I know it gets more complex now...
