### 2022-09-21 prioritizr/wdpar#61 @jeffreyhanson
Yeah that's right. The resulting dataset contains 272,466 protected areas. Yeah, I have tried running the global data with `erase_overlaps = TRUE` and it doesn't work - the geometry processing dies due to (extremely) invalid geometries and I couldn't find a work around. To address this, the package documentation recommends using `erase_overlaps = FALSE` for large datasets (eg., https://prioritizr.github.io/wdpar/articles/wdpar.html#recommended-practices-for-large-datasets, https://prioritizr.github.io/wdpar/reference/wdpa_clean.html#recommended-practices-for-large-datasets-1) and provides advice for post-processing (e.g., using `wdpa_dissolve()` to take care of overlaps).
