+++ paper/paper.bib
@@ -148,7 +148,7 @@
 
 @article{r15,
   title={{Conservation planning for adaptive and neutral evolutionary processes}},
-  author={Hanson, Jeffrey O and Marques, Adam and Ver{\'\i}ssimo, Ana and Camacho-Sanchez, Miguel and Velo-Ant{\'o}n, Guillermo and Mart{\'\i}nez-Solano, {\'I}{\~n}igo and Carvalho, Silvia B},
+  author={Hanson, J. O. and Marques, Adam and Ver{\'\i}ssimo, Ana and Camacho-Sanchez, Miguel and Velo-Ant{\'o}n, Guillermo and Mart{\'\i}nez-Solano, {\'I}{\~n}igo and Carvalho, Silvia B},
   journal={Journal of Applied Ecology},
   volume={57},
   number={11},
@@ -198,7 +198,7 @@
 }
 
 @manual{r20,
-  author = {Jeffrey O Hanson},
+  author = {Hanson, J O},
   title = {{wdpar: Interface to the World Database on Protected Areas}},
   year = {2021},
   publisher = {R package version 1.3.2},
@@ -222,4 +222,36 @@
   pages={1255--1258},
   year={2015},
   doi={10.1126/science.aac9180}
+}
+
+@manual{r23,
+  title={{ghislainv/pywdpa: pywdpa v0.1.5 (v0.1.5)}},
+  author={Vieilledent, Ghislain},
+  year={2021},
+  publisher={Zenodo},
+  doi={10.5281/zenodo.4601607}
+}
+
+@manual{r24,
+  title = {{rwdpa: World Database of Protected Areas Client}},
+  author = {Scott Chamberlain},
+  year = {2017},
+  publisher = {R package version 0.0.8.9713},
+  url = {https://github.com/ropenscilabs/rwdpa},
+}
+
+@manual{r25,
+  title = {{worldpa: An Interface to the World Database on Protected Areas (WDPA)}},
+  author = {Nicolas Casajus},
+  year = {2021},
+  publisher = {R package version 2.1.0},
+  url = {https://github.com/frbcesab/worldpa},
+}
+
+@manual{r26,
+  title={Global protected areas},
+  author={Hanson, J. O.},
+  year={2017},
+  publisher={Version 1.0.6. Zenodo},
+  doi={10.5281/zenodo.256882}
 }
