+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
 affiliations:
  - name: Department of Biology, Carleton University, Ottawa, Canada
    index: 1
-date: "08 August 2022"
+date: "21 September 2022"
 bibliography: paper.bib
 output: rticles::joss_article
 journal: JOSS
@@ -30,7 +30,7 @@ The \texttt{wdpar} \texttt{R} package provides an interface to data available on
 
 Area-based conservation measures are crucial for safeguarding biodiversity [@r5; @r6]. Examples of such measures include protected areas, marine reserves, and other effective area-based conservation measures (OECMs). Protected Planet is a key resource for area-based conservation measures, providing the World Database on Protected Areas (WDPA) and the World Database on Other Effective Area-Based Conservation Measures (WDOECM) [@r9]. These publicly available databases contain standardized data for over 270,000 protected areas and over 700 OECMs worldwide [@r9]. By detailing the designation, establishment, management, and spatial boundaries of area-based conservation measures [@r21], these databases play a vital role in monitoring and prioritizing conservation efforts [@r4; @r8].
 
-The WDPA and WDOECM require data cleaning procedures to prepare them for analysis [@r4; @r10]. These procedures include excluding areas that have yet to be fully implemented, areas that are no longer designated, and UNESCO Biosphere Reserves [@r1]. They also include geoprocessing procedures, such as repairing invalid geometries in spatial boundaries, buffering areas represented by point localities [@r3], and removing spatial overlaps [@r2]. Specifically, overlapping geometries are erased such that areas associated with more effective management categories or have historical precedence are retained. These procedures are critical to ensure accuracy in assessments of area-based conservation measures [@r1; @r2]. Yet, despite their importance, these procedures can be challenging to implement.
+The WDPA and WDOECM require data cleaning procedures to prepare them for analysis [@r4; @r10]. These procedures include excluding areas that have yet to be fully implemented, areas that are no longer designated, and UNESCO Biosphere Reserves [@r1]. They also include geoprocessing procedures, such as repairing invalid geometries in spatial boundaries, buffering areas represented by point localities [@r3], and removing spatial overlaps [@r2]. Specifically, overlapping geometries are erased such that areas associated with more effective management categories are retained [@r22] and -- in cases where geometries with the same management category overlap -- areas associated with historical precedence are retained. These procedures are critical to ensure accuracy in assessments of area-based conservation measures [@r1; @r2]. Yet, despite their importance, these procedures can be challenging to implement.
 
 The \texttt{wdpar} \texttt{R} package provides automated methods to obtain and clean the WDPA and WDOECM. The data cleaning procedures implemented in the package follow best practices [@r4; @r10] and can be performed without specialized knowledge, customized to particular use cases, and applied to the global scale. By providing this functionality, the package aims to increase accessibility to the WDPA and WDOECM.
 
@@ -39,6 +39,10 @@ The \texttt{wdpar} \texttt{R} package provides automated methods to obtain and c
 The \texttt{wdpar} \texttt{R} package is designed to provide a reproducible tool for downloading and cleaning the WDPA and WDOECM. Although the default settings for the data cleaning procedures are well-suited for national scale reporting of protected area coverage, they can be customized for other applications. For example, the precision of spatial data processing procedures can be increased so that they are suitable for local scale analyses. This is especially important because the default precision may remove smooth edges at fine scales. Additionally, the data cleaning procedures can be customized to retain protected areas regardless of their status which, in turn, could be useful for monitoring and evaluation of protected area effectiveness.
 
[…]
