### 2022-09-21 prioritizr/wdpar#62 @jeffreyhanson
Yeah, there are a few software packages that have been developed to help work with Protected Planet data. So, I've added a new paragraph to the manuscript with a brief overview of them (copied in below, and [see here for PDF](https://github.com/prioritizr/wdpar/blob/joschie-review/paper/paper.pdf)). Although DOPA have published paper documenting their methods for working with Protected Planet data (see [here](https://scholarworks.umass.edu/cgi/viewcontent.cgi?article=1017&context=foss4g)), I am unable to find any code online for these methods - so I don't think it's apropriate to include a comparison to their work because they haven't produced a software package. Is this OK?

---

> The \texttt{wdpar} \texttt{R} package provides superior functionality for processing Protected Planet data compared with other software packages. Although the \texttt{rwdpa} \texttt{R} package, \texttt{worldpa} \texttt{R} package, and the \texttt{pywdpa} \texttt{Python} package provide interfaces for downloading data from Protected Planet [r23; r24; r25], none of these software packages provide functionality for data cleaning. Additionally, the \texttt{rwdpa} \texttt{R} package has been archived. A command line tool was also developed to download and clean Protected Planet data [r26]. However, because the command line tool was implemented as a collection of \texttt{Python} scripts and configuration files, it is difficult to install and customize.
