### 2022-09-23 prioritizr/wdpar#54 @jeffreyhanson
Just to keep you in the loop, the PhantomJS web driver used by Rselenium no longer works - so the downloading functionality of the package is now broken. So I need to do some updating to wdpar so that it works again. So, I'm going to merge the PR that I've been working on with the changes to adress your feedback so I don't get into merge conflict hell. This will mean that some of your issues will be addressed, and some of them weren't. After merging the PR, I'lll then work on updating the package to fix the web driver issues, and then open a new PR to address your remaining issues. Sorry about this.
