### 2022-09-23 prioritizr/wdpar#54 @jeffreyhanson
Actually, it turns out the fix is very straightforward (replacing Rselenium with webdriver), so I'll just lump the fix into the PR.
