+++ tests/testthat/helper_skip_functions.R
@@ -4,3 +4,9 @@ skip_if_local_and_slow_internet <- function(x) {
        !identical(Sys.getenv("CI"), "true")
   testthat::skip_if(x, paste("On local and slow internet"))
 }
+
+skip_if_phantomjs_not_available <- function() {
+  testthat::skip_if_not(
+    is_phantomjs_available, "phantomJS not available"
+  )
+}
