### 2022-09-27 3241234
update paper + fix typo in docs
