+++ paper/paper.Rmd
@@ -22,13 +22,22 @@ journal: JOSS
 link-citations: yes
 ---
 
+```{r "initialization", include = FALSE}
+knitr::opts_knit$set(root.dir = normalizePath(".."))
+knitr::opts_chunk$set(
+  echo = FALSE, message = FALSE, warning = FALSE,
+  dev = "png", dev.args = list(type = "cairo"),
+  dpi = 500, fig.align = "center"
+)
+```
+
 # Summary
 
 The \texttt{wdpar} \texttt{R} package provides an interface to data available on Protected Planet (https://www.protectedplanet.net). It can be used to access the World Database on Protected Areas (WDPA) and the World Database on Other Effective Area-Based Conservation Measures (WDOECM). Additionally, it provides data cleaning procedures to prepare these databases for analysis. These data cleaning procedures are essential for ensuring correct results when using the databases. As a software package for the \texttt{R} statistical computing environment, it can easily be integrated into workflows and spatial analyses. The package has applications for conservation research. It has been used to assess performance of existing protected areas and account for such areas when identifying priority areas for conservation efforts.
 
 # Statement of need
 
-Area-based conservation measures are crucial for safeguarding biodiversity [@r5; @r6]. Examples of such measures include protected areas, marine reserves, and other effective area-based conservation measures (OECMs). Protected Planet is a key resource for area-based conservation measures, providing the World Database on Protected Areas (WDPA) and the World Database on Other Effective Area-Based Conservation Measures (WDOECM) [@r9]. These publicly available databases contain standardized data for over 270,000 protected areas and over 700 OECMs worldwide [@r9]. By detailing the designation, establishment, management, and spatial boundaries of area-based conservation measures [@r21], these databases play a vital role in monitoring and prioritizing conservation efforts [@r4; @r8].
+Area-based conservation measures are crucial for safeguarding biodiversity [@r5; @r6]. Examples of such measures include protected areas, marine reserves, and other effective area-based conservation measures (OECMs). Protected Planet is a key resource for area-based conservation measures, providing the World Database on Protected Areas (WDPA) and the World Database on Other Effective Area-Based Conservation Measures (WDOECM) [@r9]. These publicly available databases contain standardized data for over 270,000 protected areas and over 800 OECMs worldwide [@r9]. By detailing the designation, establishment, management, and spatial boundaries of area-based conservation measures [@r21], these databases play a vital role in monitoring and prioritizing conservation efforts [@r4; @r8].
 
 The WDPA and WDOECM require data cleaning procedures to prepare them for analysis [@r4; @r10]. These procedures include excluding areas that have yet to be fully implemented, areas that are no longer designated, and UNESCO Biosphere Reserves [@r1]. They also include geoprocessing procedures, such as repairing invalid geometries in spatial boundaries, buffering areas represented by point localities [@r3], and removing spatial overlaps [@r2]. Specifically, overlapping geometries are erased such that areas associated with more effective management categories are retained [@r22] and -- in cases where geometries with the same management category overlap -- areas associated with historical precedence are retained. These procedures are critical to ensure accuracy in assessments of area-based conservation measures [@r1; @r2]. Yet, despite their importance, these procedures can be challenging to implement.
 
@@ -43,6 +52,18 @@ The \texttt{wdpar} \texttt{R} package has several applications for conservation 
 # Comparison with other software packages
 
[…]
