+++ paper/paper.bib
@@ -93,8 +93,8 @@
   author={UNEP-WCMC and IUCN},
   year={2022},
   title={{Protected Planet: The World Database on Protected Areas (WDPA) and World Database on Other Effective Area-based Conservation Measures (WD-OECM)  [Online]}},
-  note={May 2022},
-  publisher={{May 2022, Cambridge, UK: UNEP-WCMC and IUCN}},
+  note={September 2022},
+  publisher={{September 2022, Cambridge, UK: UNEP-WCMC and IUCN}},
   url={www.protectedplanet.net}
 }
 
@@ -173,7 +173,8 @@
   author={Gonzalez, Baltazar and Brook, Federico and Martin, Gabriel M},
   journal={Hystrix, the Italian Journal of Mammalogy},
   year={2022},
-  volume={In press},
+  volume={33},
+  pages={0},
   doi={10.4404/hystrix-00489-2021},
 }
 
@@ -254,4 +255,16 @@
   year={2017},
   publisher={Version 1.0.6. Zenodo},
   doi={10.5281/zenodo.256882}
+}
+
+@article{r28,
+  author = {Edzer Pebesma},
+  title = {{Simple features for R: Standardized support for spatial vector data}},
+  year = {2018},
+  journal = {{The R Journal}},
+  doi = {10.32614/RJ-2018-009},
+  url = {https://doi.org/10.32614/RJ-2018-009},
+  pages = {439--446},
+  volume = {10},
+  number = {1},
 }
