### 2022-09-27 prioritizr/wdpar#54 @jeffreyhanson
@Jo-Schie, I've just pushed a commit to PR #57 so that the manuscript now contains a short case study and a map showing some protected areas. When you get a chance, could you please take a look? This PR also contains updates to address the web driver data download issues too. To help make it easier to review, I've attached the PDF for the updated manuscript to here: [paper.pdf](https://github.com/prioritizr/wdpar/files/9660127/paper.pdf). If the case study is satisfactory, my understanding is that merging the PR would address all your concerns - is that right? I just want to make sure I haven't missed anything.
