+++ .github/workflows/documentation.yaml
@@ -55,6 +55,7 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 4
+          pak-version: devel
           extra-packages: |
             any::rcmdcheck
             any::urlchecker
