### 2022-09-27 ba0a43d
try using stable pak version
