+++ .github/workflows/documentation.yaml
@@ -55,7 +55,6 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 4
-          pak-version: devel
           extra-packages: |
             any::rcmdcheck
             any::urlchecker
