+++ .github/workflows/documentation.yaml
@@ -44,29 +44,17 @@ jobs:
           r-version: ${{ matrix.config.r }}
           use-public-rspm: true
 
-      - name: Install system dependencies
-        run: |
-          sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
-          sudo apt update
-          sudo apt-get install \
-            libgdal-dev libgeos-dev libproj-dev \
-            libgmp3-dev libmpfr-dev libudunits2-dev
-
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 4
           extra-packages: |
             any::rcmdcheck
             any::urlchecker
+            any::remotes
             cran/XML
-            prepr=?ignore
+            dickoa/prepr
           needs: |
             check
-
-      - name: Install spatial R packages
-        run: |
-          pak::pkg_install("dickoa/prepr")
-        shell: Rscript {0}
 
       - name: Session information
         run: |
