+++ .github/workflows/R-CMD-check-windows.yaml
@@ -56,7 +56,7 @@ jobs:
           cache-version: 4
           extra-packages: |
             any::rcmdcheck
-            cran/XML
+            any::XML
             prepr=?ignore
           needs: |
             check
