### 2022-09-28 prioritizr/wdpar#54 @Jo-Schie
I guess you could also Show also some Areal statistics eg sqkm in each iucn category to show the purpose of cleaning the dataset... but that's up to you. @jeffreyhanson  i will close This issue so you can merge your changes.
