### 2022-09-28 prioritizr/wdpar#54 @jeffreyhanson
Thanks! Yeah, I wanted to keep the case study simple. Although I like the idea of providing aerial statistics, I worry that this is difficult to do without adding too much extra code that would require explanation (e.g., calculating aerial statistics would probably be best done using dplyr + sf). I suppose there are base package alternatives (e.g., using base::aggregate), but I wouldn't want to recommend these as best practice (and tidyverse users might find this concercing/confusing). Since I can't think of a simple way to do this, I'll just keep the current stats.
