+++ .github/workflows/R-CMD-check-ubuntu.yaml
@@ -56,7 +56,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          cache-version: 4
+          cache-version: 5
           extra-packages: |
             any::rcmdcheck
             any::covr
