+++ NEWS.md
@@ -1,12 +1,13 @@
 # wdpar 1.3.3.1
 
-- Update paper for JOSS submission (#53, #62).
+- Update paper for JOSS submission (#53, #54, #62).
 - Update vignette with new section on local scale analyses (#53).
 - Update `wdpa_fetch()` function to use the webdriver package for obtaining data
   (replacing Rselenium package as a dependency) (#63).
 - Update `st_repair_geometry()` to be more robust.
 - Fix failing tests for `st_repair_geometry()` function.
 - Update documentation for `wdpa_clean()` function.
+- Fix broken URL in vignette.
 
 # wdpar 1.3.3
 
