### 2022-09-28 d075533
address jo-schie's comments (#57)
- Update paper for JOSS submission (fix #53, fix #54, fix #62).
- Update vignette with new section on local scale analyses (fix #53).
- Update `wdpa_fetch()` function to use the webdriver package for obtaining data  (replacing Rselenium package as a dependency) (fix #63).
- Update `st_repair_geometry()` to be more robust.
- Fix failing tests for `st_repair_geometry()` function.
- Update documentation for `wdpa_clean()` function.
- Fix broken URL in vignette.
- Fix CI.
