+++ R/st_repair_geometry.R
@@ -67,6 +67,23 @@ st_repair_geometry <- function(x, geometry_precision = 1500) {
   # extract polygons and points (if needed)
   x2 <- extract_polygons_and_points(x2)
 
+  # dissolve by repair id
+  if (!identical(anyDuplicated(x2[["_repair_id"]]), 0L)) {
+    x2 <- split(x2, x2[["_repair_id"]])
+    x2_df <- tibble::tibble(`_repair_id` = as.integer(names(x2)))
+    x2 <- lapply(x2, sf::st_geometry)
+    x2 <- lapply(x2, sf::st_union)
+    x2 <- do.call(c, x2)
+    x_df <- match(x2_df[["_repair_id"]], x[["_repair_id"]])
+    x_df <- sf::st_drop_geometry(x)[x_df, , drop = FALSE]
+    x_df <- x_df[, setdiff(names(x_df), "_repair_id"), drop = FALSE]
+    x2_df <- tibble::as_tibble(cbind(x2_df, x_df))
+    x2_df <- x2_df[, names(sf::st_drop_geometry(x)), , drop = FALSE]
+    x2_df$geometry <- x2
+    x2 <- sf::st_sf(x2_df)
+    rm(x_df, x2_df)
+  }
+
   # detect if any invalid geometries persist
   ## subset repaired polygons
   x_sub <- x[match(x2[["_repair_id"]], x[["_repair_id"]]), , drop = FALSE]
