### 2022-09-28 c83514c
update warning + docs on versions, #59
