### 2022-09-28 prioritizr/wdpar#59 @jeffreyhanson
@DrMattG, I've just created a PR to address this issue. When you get a chance, could you please take a look? I've put the changes specifically related to this issue in this commit: https://github.com/prioritizr/wdpar/pull/64/commits/c83514c2fe4ef1bcc325549270f52975c07f4500.
