### 2022-09-28 bb04582
skip dois in url checks
