+++ .github/workflows/documentation.yaml
@@ -83,6 +83,7 @@ jobs:
         if: runner.os == 'Linux' && matrix.config.r == 'release'
         run: |
           result <- urlchecker::url_check()
+          result <- result[!startsWith(result$URL, "https://doi.org/"), , drop = FALSE]
           if (nrow(result) > 0) {
             print(result)
             stop("Invalid URLs detected")
