### 2022-10-03 81b47ae
address reviewer comments by @drmattg (#64)
* update README, fix #58
* update warning + docs on versions, #59
* update pkg docs
* skip DOIs in url checks
