+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: wdpar
 Type: Package
-Version: 1.3.3.1
+Version: 1.3.3.2
 Title: Interface to the World Database on Protected Areas
 Description: Fetch and clean data from the World Database on Protected
     Areas (WDPA) and the World Database on Other Effective Area-Based
