### 2022-10-03 prioritizr/wdpar#59 @jeffreyhanson
Woops, forgot tag this issue in the commit when merging this PR, so I'll close this manually.
