### 2022-10-03 #4594 @jeffreyhanson
Massive thank you to @Jo-Schie and @DrMattG for making the time to review this submission. Their feedback was extremely helpful for improving the package and manuscript! I think I've fully addressed all of their comments now - but please correct me if I've got this wrong - so what should I do next?
