### 2022-10-12 #4594 @jeffreyhanson
@Jo-Schie, I just wanted to follow up and check if your concerns about Performance and State of the Field were addressed? My undestanding is that you raised these concerns in https://github.com/prioritizr/wdpar/issues/61 and https://github.com/prioritizr/wdpar/issues/62 (respectively) and I addressed these concerns by providing a log file with timings and updating the manuscript?
