### 2022-10-13 #4594 @Jo-Schie
They were both adequately addressed during the review. I marked the checkboxes accordingly.
