### 2022-10-13 #4594 @jeffreyhanson
Awesome! @Jo-Schie, could you please mark the checkbox for your review on this thread too (https://github.com/openjournals/joss-reviews/issues/4594#issuecomment-1193077208)? On my screen, the "State of the Field" checkbox isn't marked. Based on @martinfleis previous comment, I think they might need this to confirm that your review is complete (please correct me if I've got this wrong)?
