### 2022-10-13 #4594 @martinfleis
Thank you all! I made sure all is checked as is supposed to based on the discussions.

@jeffreyhanson The submission is now almost ready to be published.

The next steps you need to do now:

- [ ] Make a new tagged release (or let me know if the latest one contains all the changes) of the software, and list the version tag of the archived version here.
- [ ]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper.
