### 2022-10-13 #4594 @jeffreyhanson
Awesome - thanks so much @martinfleis, @DrMattG, and @Jo-Schie!

@martinfleis, I've created a tagged release (see https://github.com/prioritizr/wdpar/releases/tag/v1.3.3.2) and archived this on Zenodo (see https://doi.org/10.5281/zenodo.7195676). I also have edited the metadata so that the author list and title are identical to that of the paper. Please let me know if there's anything else I can do top help?
