### 2022-10-15 6956834
fix phantomjs install in tests
