+++ tests/testthat/test_000_setup.R
@@ -5,7 +5,7 @@ test_that("phantomjs", {
   skip_on_cran()
   # check if phantomjs installed and install if needed
   if (!isTRUE(wdpar:::has_phantomjs())) {
-    phantomjs::install_phantomjs()
+    webdriver::install_phantomjs()
   }
   # update variable
   is_phantomjs_available <<- wdpar:::has_phantomjs()
