+++ vignettes/wdpar.Rmd
@@ -291,7 +291,7 @@ Although the [World Database on Protected Areas (WDPA)](https://www.protectedpla
     + [Collaborative Australian Protected Area Database (CAPAD)](https://www.dcceew.gov.au/environment/land/nrs/science/capad)
     + [Conservation Areas Reporting and Tracking System (CARTS)](https://ccea-ccae.org)
 * The United States of America
-    + [Protected Areas Database of the US, PAD-US (CBI Edition)](https://consbio.org/products/projects/pad-us-cbi-edition)
+    + [Protected Areas Database of the US (PAD-US)](https://doi.org/10.5066/P9Q9LQ4B)
 
 ### Citation
 
