### 2022-10-23 #4594 @jeffreyhanson
@martinfleis, just to follow up, I think I've completed all those steps (see post above for release and DOI) - is there anything else I need to do?
